fix: itest: check for closed connection
