+++ lib/retry/retry.go
@@ -1,24 +1,14 @@
 package retry
 
 import (
-	"errors"
-	"reflect"
 	"time"
 
 	logging "github.com/ipfs/go-log/v2"
+
+	"github.com/filecoin-project/lotus/api"
 )
 
 var log = logging.Logger("retry")
-
-func ErrorIsIn(err error, errorTypes []error) bool {
-	for _, etype := range errorTypes {
-		tmp := reflect.New(reflect.PointerTo(reflect.ValueOf(etype).Elem().Type())).Interface()
-		if errors.As(err, tmp) {
-			return true
-		}
-	}
-	return false
-}
 
 func Retry[T any](attempts int, initialBackoff time.Duration, errorTypes []error, f func() (T, error)) (result T, err error) {
 	for i := 0; i < attempts; i++ {
@@ -28,7 +18,7 @@ func Retry[T any](attempts int, initialBackoff time.Duration, errorTypes []error
 			initialBackoff *= 2
 		}
 		result, err = f()
-		if err == nil || !ErrorIsIn(err, errorTypes) {
+		if err == nil || !api.ErrorIsIn(err, errorTypes) {
 			return result, err
 		}
 	}
